Move subtree and forest to dataset; Move TailSkipForSIMD; lint change
