+++ cpp/src/arrow/dataset/CMakeLists.txt
@@ -153,6 +153,7 @@ add_arrow_dataset_test(file_ipc_test)
 add_arrow_dataset_test(file_test)
 add_arrow_dataset_test(partition_test)
 add_arrow_dataset_test(scanner_test)
+add_arrow_dataset_test(subtree_test)
 
 if(ARROW_CSV)
   add_arrow_dataset_test(file_csv_test)
